Libera: Fix long standing EPICS update bug
For a long time the Libera EPICS driver has suffered from a bug where
updates to menu settings aren't always reflected over channel access.
It turns out that the correct solution is to initialise the mlst field
at the same time as setting the VAL field of every out record.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/ioc/Libera@55892 e099a375-04f9-0310-9d5f-a741eaff62e1