Libera: Fix possible cryptic problem with restart
We need to use vfork() when restarting EPICS because otherwise when memory
is very commited the restart can fail.  Unfortunately the rules of what
can be done in the child before calling _exit or execve are rather strict.
This commit is a rewrite to simplify the execution of this code.

Also sort out the handling of stdin in daemon mode, rather than closing stdin
instead we redirect /dev/null to stdin, this is done in the shell script.
This is really part of this, as the handling of stdin affects children.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/ioc/Libera@65908 e099a375-04f9-0310-9d5f-a741eaff62e1